improvements to the PUT method used to update registered persons

+++ eventman_server.py
@@ -35,6 +35,7 @@ class BaseHandler(tornado.web.RequestHandler):
     _bool_convert = {
         '0': False,
         'n': False,
+        'f': False,
         'no': False,
         'off': False,
         'false': False
@@ -74,6 +75,7 @@ class CollectionHandler(BaseHandler):
     @gen.coroutine
     def get(self, id_=None, resource=None, resource_id=None, **kwargs):
         if resource:
+            # Handle access to sub-resources.
             method = getattr(self, 'handle_get_%s' % resource, None)
             if method and callable(method):
                 self.write(method(id_, resource_id, **kwargs))
@@ -92,6 +94,7 @@ class CollectionHandler(BaseHandler):
     def post(self, id_=None, resource=None, resource_id=None, **kwargs):
         data = escape.json_decode(self.request.body or '{}')
         if resource:
+            # Handle access to sub-resources.
             method = getattr(self, 'handle_%s_%s' % (self.request.method.lower(), resource), None)
             if method and callable(method):
                 self.write(method(id_, resource_id, data, **kwargs))
@@ -108,11 +111,14 @@ class CollectionHandler(BaseHandler):
     @gen.coroutine
     def delete(self, id_=None, resource=None, resource_id=None, **kwargs):
         if resource:
+            # Handle access to sub-resources.
             method = getattr(self, 'handle_delete_%s' % resource, None)
             if method and callable(method):
                 self.write(method(id_, resource_id, **kwargs))
                 return
-        self.db.delete(self.collection, id_)
+        if id_:
+            self.db.delete(self.collection, id_)
+        self.write({'success': True})
 
 
 class PersonsHandler(CollectionHandler):
@@ -121,6 +127,14 @@ class PersonsHandler(CollectionHandler):
     object_id = 'person_id'
 
     def handle_get_events(self, id_, resource_id=None, **kwargs):
+        # Get a list of events attended by this person.
+        # Inside the data of each event, a 'person_data' dictionary is
+        # created, duplicating the entry for the current person (so that
+        # there's no need to parse the 'persons' list on the client).
+        #
+        # If resource_id is given, only the specified event is considered.
+        #
+        # If the 'all' parameter is given, every event (also unattended ones) is returned.
         args = self.request.arguments
         query = {}
         if id_ and not self.tobool(args.get('all')):
@@ -136,6 +150,8 @@ class PersonsHandler(CollectionHandler):
                     person_data = persons
                     break
             event['person_data'] = person_data
+        if resource_id and events:
+            return events[0]
         return {'events': events}
 
 
@@ -145,15 +161,20 @@ class EventsHandler(CollectionHandler):
     object_id = 'event_id'
 
     def handle_get_persons(self, id_, resource_id=None):
+        # Return every person registered at this event, or the information
+        # about a specific person.
         query = {'_id': id_}
         event = self.db.query('events', query)[0]
         if resource_id:
             for person in event.get('persons', []):
                 if str(person.get('person_id')) == resource_id:
                     return {'person': person}
-        return {'persons': event.get('persons') or {}}
+        if resource_id:
+            return {'person': {}}
+        return {'persons': event.get('persons') or []}
 
     def handle_post_persons(self, id_, person_id, data):
+        # Add a person to the list of persons registered at this event.
         doc = self.db.query('events',
                 {'_id': id_, 'persons.person_id': person_id})
         if '_id' in data:
@@ -162,21 +183,23 @@ class EventsHandler(CollectionHandler):
             merged, doc = self.db.update('events',
                     {'_id': id_},
                     {'persons': data},
-                    operator='$push',
+                    operation='append',
                     create=False)
         return {'event': doc}
 
     def handle_put_persons(self, id_, person_id, data):
+        # Update an existing entry for a person registered at this event.
         merged, doc = self.db.update('events',
                 {'_id': id_, 'persons.person_id': person_id},
-                data, create=False)
+                data, updateList='persons', create=False)
         return {'event': doc}
 
     def handle_delete_persons(self, id_, person_id):
+        # Remove a specific person from the list of persons registered at this event.
         merged, doc = self.db.update('events',
                 {'_id': id_},
                 {'persons': {'person_id': person_id}},
-                operator='$pull',
+                operation='delete',
                 create=False)
         return {'event': doc}
 
@@ -201,7 +224,7 @@ class EbCSVImportPersonsHandler(BaseHandler):
         'ID ordine': 'order_nr',
         'Prefisso (Sig., Sig.ra, ecc.)': 'name_title',
     }
-    keepPersonData = ('name', 'surname', 'email')
+    keepPersonData = ('name', 'surname', 'email', 'name_title')
 
     @gen.coroutine
     def post(self, **kwargs):
@@ -237,7 +260,7 @@ class EbCSVImportPersonsHandler(BaseHandler):
                                 {'_id': event_id, 'persons.person_id': person_id}):
                             self.db.update('events', {'_id': event_id},
                                     {'persons': person},
-                                    operator='$addToSet')
+                                    operation='appendUnique')
                             reply['new_in_event'] += 1
         self.write(reply)
